requirements.txt
# Core AI and Image Processing
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.21.0
transformers>=4.30.0
accelerate>=0.21.0
xformers>=0.0.20

# Image Processing and Computer Vision
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.11.0
scikit-image>=0.21.0

# Face Detection and Enhancement
mediapipe>=0.10.0
dlib>=19.24.0
face-recognition>=1.3.0
insightface>=0.7.3

# Background Removal
rembg>=2.0.50
backgroundremover>=0.2.8

# API and Web Interface
fastapi>=0.100.0
uvicorn>=0.23.0
streamlit>=1.25.0
gradio>=3.40.0

# Utilities
requests>=2.31.0
tqdm>=4.65.0
matplotlib>=3.7.0
seaborn>=0.12.0
python-dotenv>=1.0.0
